Compile in release mode on Windows
This will avoid errors like the following on Windows:
"The code execution cannot proceed because VCRUNTIME140_1D.dll was not found.
Reinstalling the program may fix this problem."
Even though it would be convenient to provide a debug version to help us
debugging issues in tango-controls, this is not convenient if the user cannot
execute UniversalTest out of the box without having to install extra dlls.
As explained in https://github.com/conda/conda-verify/issues/67, most users do
not have the debug dlls installed.